requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
nbformat==4.4.0
numpy==1.15.0
pandas==0.23.3
plotly==3.1.0
psycopg2==2.7.5
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.10
traitlets==4.3.2
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
